requirements.txt
# Core Simulation & Data
mesa>=0.8.10
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=5.4.0

# Visualization & Plotting
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Dashboard (Solara)
solara>=1.0.0
solara-plotly>=0.1.0

# Data Processing
scipy>=1.7.0

# Jupyter & Notebooks
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0

# Optional: Advanced Analytics
scikit-learn>=0.24.0
statsmodels>=0.13.0

# Optional: Database support
sqlalchemy>=1.4.0

# Development & Testing
pytest>=6.2.0
black>=21.0
flake8>=3.9.0
